Added limited decompilation
Disassembling is the prefered way to use the script, but I added a decompilation (merging instructions into function calls and some other patterns)
Disassembling and reassemling of all scena files give the same binary than the original scripts, except:
a0000
c0400
sys_event
m3100

However the difference is minor and I am not worried; I think some of them are wrongly generated (C0400 has a POP after an EXIT(), which is super weird). A small fix can be applied to the script to make it work though.