Reversed the parameters order, fixed a few problems
notably, Return() gives a better POP() quantity at the end (instead of removing the number of input parameters, it just clears the stack. This seems to work well EXCEPT for particular functions (mainly in ani files?) with a particular patterns, where I have no idea how the POP is calculated. This honestly requires further investigation but I had problems debugging the game so it's probably put on hold

The whole code is even messier than before honestly, it's a nightmare.

The only file I think I can't recompile (using the decompiled version, not the disassembled one) right now except the obviously broken ones is chr5007. This one might have the particular pattern talked above so it might be a matter of solving it once and for all.

Still debating whether the decompiled version or the disassembled version should be the default mode since disassembling is way safer than decompiling regarding the data integrity. Anyway fuck that tool